requirements.txt
kedro~=1.3.1
kedro-datasets[pandas,pickle]>=3.0
catboost>=1.2
scikit-learn>=1.3
pandas>=2.0
pyarrow>=14.0
numpy>=1.26
dash>=2.17
dash-bootstrap-components>=1.6
plotly>=5.20
pyyaml>=6.0
ipython>=8.10
jupyterlab>=3.0
notebook
